Use DraftVaccinationRecord
This updates how vaccination recording and editing is build to use the
new `DraftVaccinationRecord` class instead of storing draft records in
the database. This implements the behaviour in the prototype where users
can make numerous changes before committing the changes.

This will also result in there being no more draft records in the
database so a follow up change will be made to remove them.