Use DraftGillickAssessment
This updates how gillick assessment recording and editing is built to
use the new `DraftGillickAssessment` class instead of storing draft
records in the database. This implements the behaviour in the
prototype where users can make numerous changes before committing the
changes.

This will also result in there being no more draft records in the
database so a follow up change will be made to remove them.