Fix issues count in the secondary nav under Programmes
The import issues counter in the secondary nav was double counting import issues where both the vaccination record and the patient have pending changes.

Because the import issues controller is doing this correctly, there was also a discrepancy between the count and what the user saw when they navigated to the import issues list.