Add TIME_OF_VACCINATION immunisation import column
This adds a new column that we can optionally parse when importing
immunisations to allow the user to provide the time of vaccination. This
will mostly be used for offline recording, but could also be used when
importing historical records.

To account for different ways of formatting times we parse a number of
different formats, but in the documentation I've only written one
explicit format that's the most unambigious.