Add PERFORMING_PROFESSIONAL_EMAIL immunisation import column
This adds a new optional column that can be parsed allowing the uploader
to specify the email address of the user who performed the vaccination.
This allows us to create a direct link in our database between
vaccinations and users rather than filling in the free text family/given
name fields.