Prevent creation of session dates in current academic year
When importing immunisations, we now validate that the date of
vaccination matches an existing session date if we're importing records
for the current academic year rather than creating new session dates.

To support this the `location` and `school` method can now be used even
if the row doesn't validate (as they're used in this validation).