Automatically pick vaccine if one exists
When changing the outcome of a vaccination record, if changing to
vaccinated, we can automatically choose the vaccine for the user if
there's only one possible option which saves time.

It also fixes an issue we have for HPV where the batches and delivery
site/methods are unavailable, however this problem will come again for
Flu where there are multiple vaccines.