Add CLINIC_NAME column
This adds a new column which we parse when when importing immunisations
for clinics which currently doesn't allow for a way of specifying the
name of the clinic, unlike when using the app directly record
vaccinations. This helps us to support offline recording of vaccinations
but is unlikely to be used for historical imports.