Remove FLIPPER_INITIALIZERS for basic_auth
This ensures that the first time an environment is deployed, it's behind
basic auth.

We then override this on some of our environments, like training.

However, if you then:

- Clear the database
- Redeploy at some point

The flag will get re-initialised to `true` and users will lose access.

Since this has happened twice on the training environment, this
initializer is doing more harm than good.