Add ability to delete vaccination records
This adds the basic functionality that allows users to delete
vaccination records for sessions that are still open. There's no design
for this in the prototype so the interface might need changing in the
future.

Some further improvements required include soft-deletion so old
vaccination records can be shown in the activity log and a feature test
to ensure the records can only be deleted for open sessions.