Only show kept vaccination records
This applies a default scope to all vaccination record relationships and
the vaccination record policy to only return kept records. This makes it
easy to keep our existing code and avoid showing discarded records, but
this approach is perhaps not the best practice as it means we've got
implicit defaults. We may find that we need to reverse this decision in
the future, but for now I think this is the quickest and safest
approach.