Generate the kid for the JWT used with NHS API
We had this hard-coded as "mavis-int-1", which follows the guidelines
laid out by the NHS for what kid to use. See:

https://digital.nhs.uk/developer/guides-and-documentation/security-and-authorisation/application-restricted-restful-apis-signed-jwt-authentication#step-2-generate-a-key-pair

To different keys in different envs we need a way to link the kid to the
private keys we use, either by pre-generating them and storing them as a
jwk, or by using the ruby-jwt gem to do derive it from our private key.
The latter approach has been chosen here.

The change to the test credentials file adds a "jwt_private_key" to use
in our tests.