Update NHS API private key for INT env
Because we've changed how we generate the kid (it used to be statically
set to "mavis-int-1", see previous commit), we will need to ask NHS API
to update the JWKS they are hosting for us. This will probably go
smoother if we provide them with a new public key at the same time.

We must make sure NHS API has updated their self-hosted JWKS before we
deploy this change.