Rename VaccinationRecord#administered_at
To represent the fact that this field no longer represents only if the
vaccination record was administered but more generally the date at which
the vaccination record was performed (with either an administered
outcome or something else).

I've also refactored all the places where we were previously using the
`created_at` value where instead we should have been using a performed
date.