Ignore invalidated triages
When looking at the triages for a patient session we should ignore any
triages that have been invalidated. There isn't currently a design for
how these should be shown, so the quicker implementation is to simply
hide these triages, even though we have a record of them stored in the
database.