Invalidate triage after consent
If consent is withdrawn or invalidated, we need to invalidate any
triages that were performed on the patient using the old consent
responses. Currently this invalidates all triages, which works in most
situations, however could be unnecessary if, for example, both parents
consent and then one parent withdraws their consent, in that case you'd
still expect the patient to be safe to vaccinate. However this seemed
like a simpler solution for now, especially while we don't surface the
invalidated triages, and we can iterate on this in the future.