Attach NHS number to consent form when matching
If we find a PDS patient with an NHS number that doesn't exist in our
database, and no patients match with the consent form, we should update
the consent form with the PDS NHS number. This will allow nurses to
create a patient record form this consent form at a later stage.