Ensure self-consent triages are invalidated
This was missed in d62ed3d278f3224f1fe43cfce2c88d5ab5bcb04e to cover the
situation where a session has been closed, self-consents have been
invalidated and therefore any associated triages must be invalidated
too.

If we encounter a similar bug it might be worth refactoring this code to
handle the invalidation in the consent model rather than in the
controllers.

I also think it's worth adding a feature test for this in the future.