Remove pending school_id if session is proposed
We have two different mechanisms that are both handling similar
functionality:

- The `patient.pending_changes` hash can contain a `school_id` which can
  be applied to update their location
- The `patient_session.proposed_session` id that can be applied to
  update a patient's session and additionally their location

The `proposed_session` mechanism is preferable in situations where the
only thing that's changed as the result of a class list import is the
patient's school. The patient shouldn't appear in the list of duplicate
records to review, and instead appear in the session moves UI.

To prevent them from appearing in the duplicate records list, we can
remove the `pending_changes["school_id"]` if a session is successfully
proposed on their existing patient sessions.