Use DraftConsent
This updates how consent recording and editing is built to use the new
`DraftConsent` class instead of storing draft records in the database.
This implements the behaviour in the prototype where users can make
numerous changes before committing the changes. It also fixes a number
of bugs, notable where parents and triages were being created before the
user was confirming the consent.

This will also result in there being no more draft records in the
database so a follow up change will be made to remove them.