Show patient address when comparing duplicates
Various parts of the patient's address can change when importing
from class lists and cohort lists so we should ensure this is all
visible to the user rather than just showing the postcode.

In all cases where the patient summary component is used we are now
showing the full address and preferred name, so these options can be
removed and we can show it in all cases.