Remove draft parents
We no longer need the concept of draft parents as we record consent in
the browser session and only commit to the database after the confirm
step. This helps to clean up the data in our database and avoid draft
records accidentally being exposed or simply existing with no purpose.