Allow downloading and uploading vaccinations without postcodes
This is necessary to support restricted patients, ones where the
postcode is not allowed to be seen by the nurse. We need to update the
offline recording spreadsheet to not include the postcode for any
restricted patients, and similarly, we need to make it possible to
upload a vaccination file without a postcode.