Model Gillick assessments as one-to-many
This better represents their real database structure where the foreign
key exists on the Gillick assessment model. It'll also allow us to
modify the editing process to be amend-only, meaning that old Gillick
assessments are kept around but only the latest one is shown to nurses.
This makes it possible to show all Gillick assessments in the activity
log for a patient.