Show archived or expired batches when editing records
When editing a vaccination record, it's possible that the batch has
expired or been archived since the vaccination record was created. In
that case we should still show this batch as an option if the user
chooses to edit the record.

I've also refactored the policies and controllers to simplify the code
slightly.