Refactor deployment workflow
We have 3 `deploy*.yml` that all do the same thing roughly.

`test` and `pentest` are deployed to continuously, but we'd like
`accessibility` and `training` to be as well.

`production` needs to use a separate AWS account ID and only be manually
triggered.

This extracts a reusable `_deploy-template.yml` and configures
`deploy.yml` to deploy the four staging environments in parallel using a
`matrix` strategy.

`production` can still be manually triggered using the interface.