Allow navigating back to draft consent/vaccination record
We previously used to reset the session after the user confirms
changes, this was mostly to handle the limited size of the session
when it was stored in cookies. However, now that we're storing these in
the database we can leave the session and allow the user to go back
without raising an error.