Override load balancer names
By default copilot (or AWS) creates our load balancers with a name that
has a double-dash. This is causing issues with NHS's DNS provider, their
systems isn't able to setup a DNS record with a double-dash in it,
supposedly citing RFC5891, which doesn't actually apply to these names.

This change sets the name of the load balancers to a name without a
double-dash, and also with the environment name in it which fixes a
small niggle we had when viewing these resources in AWS where it was
difficult to know which environment a load balancer belonged to.