Rename import recorded_at column
To `processed_at`, and rename other related methods. This is just to
clarify the meaning of this column in relation to the `recorded_at`
column we've got left on consent forms. I propose we switch to the name
`processed_at` which matches the phrasing we show to the user anyway,
with the "Processing" tag and the "Import processing start" flash
message.

We use the term `recorded_at` to refer to draft records once they've
been confirmed, so I wanted us avoid using the same name to refer to
two different concepts.

Imports used to happen in two stages, first we would add draft versions
of all the records and then the user would confirm and then they would
be recorded. This is how we've ended up with a `recorded_at` column,
but we no longer do imports in two steps, so I think we can go back to
calling it `processed_at`.