Refactor locations factories
This changes the factories for locations to use separate factories for
the three types of locations, schools, community clinics and generic
clinics. This makes it impossible to use the school traits in clinics
and vice-versa, by introducing a clear separation between the types of
locations.

I also think this helps to make the code that uses these factories
clearer and easier to read.