Don't move patient if sessions are the same
If we're trying to move a patient between the same session we don't need
to propose a change since there's no move to take place. This also fixes
a bug where confirming the move would remove the patient from the
session because we end up with a single patient session that ends up
being deleted by `confirm!` rather than two patient sessions.

I've also improved our clinic parental consent feature test to ensure
this scenario is fully tested, including the nurse logging in and
checking for any moves and the patient school details.