Add school move confirmation page
This changes how school moves are confirmed or ignored to add a new page
which gives the user the option to choose from radio buttons on what
action to take.

This will be extended in the future to allow the user to choose whether
to keep the patient in the community clinic or not.