Allow moving community clinic patients to school
This adds an option when confirming a school move that allows the nurses
to confirm whether to keep the patient in the community clinic or move
them to an upcoming school session if that's available.

This is necessary to handle the situation where a parent fills in the
consent form for a community clinic and tells us that their child
belongs to a school, however we can't move them to the school in case
the parent has already booked them in to a clinic.