Refactor OfflineSessionExporter
This refactors the offline session exporter to move the types and styles
on a `Cell` class, meaning they can be controlled at an individual cell
level, while also keeping a cache of styles to ensure the same style
isn't created multiple times in the same sheet.

This change will also make it easier to implement row-level styles, such
as coloured backgrounds which we need for patients without consent.