Avoid creating duplicate gillick assessments
If the user submits the Gillick assessment form with no changes, we want
to avoid creating unnecessary duplicate records (which also leads to
them showing in the activity log).

This was supposed to be working before but wasn't because `dup` creates
a duplicate record with changes so the new record was always being
saved.