Use patient sessions in children view
This fixes an issue where the status wasn't filtering because it was
showing a list of patients which don't have a status. We had previously
changed this to show patients who aren't in sessions, but because of the
new school moves work, a patient should always belong to at least one
session, and if not then a bug has happened elsewhere in the service.