Replace /ping with Rails 7.2 default /up
New Rails applications can rely on the built-in `rails/health#show`
endpoint for the healthcheck.

Rails 8.0 applications can additionally use the
`silence_healthcheck_path` config option, to replace our custom
initialiser.

While we are not on Rails 8, so the config is currently a no-op, we will
migrate at some point and that should straighten itself out.

The reason we want to silence the health endpoint is because we rely on
pinging it for the offline working scenarios, but those are not in use
currently.