Fix bug related to PatientSession#no_consent?
This method is used in a number of places to determine the status of a
patient session. In the case where one consent was invalidated and one
consent was not provided, this method was returning `false` which would
then lead to consent being shown as given when this was not the case.

I've added some unit tests for this method to make sure we cover this
scenario.