Refactor vaccination flow to match prototype
This changes the vaccination form shown at the bottom of the patient
page to ask only whether the patient is ready to be vaccinated and then
ask the user to specify the delivery location after choosing an option.

This makes the necessary changes in preparation for adding the
new pre-screening questions.