Refactor draft model back links
This changes the back links used in the consent and vaccination record
wizards, mostly to fix a bug where clicking back on the first page of
the wizard would reload the page. This was happening because the
`current_page?` method doesn't handle the URL translations correctly.
For vaccination records, the page was `"date-and-time"` and the first
wizard step is `:date_and_time` and the comparison was failing.