Add a default name to AppBacklinkComponent
This component is often used as part of a wizard where we've got
"previous page" hard coded as the name of the link, so instead we can
move this in to the component itself to avoid repeating the string in
lots of places.

I've also changed the first argument to not be a keyword argument as
it's likely to be obvious that it's the path of the backlink, since it's
the only required argument.