Add AppLogEventComponent
This adds a component which renders a single log event entry, containing
a title, a body, and a subtitle (combination of a date/time and an
user). There are two places in the service currently where we render
elements like this: the activity log and the triage notes.

This also refactors those two components to use the
`AppLogEventComponent` to avoid duplication and inconsistencies.