Record access log entries
When viewing the page for an individual page, or looking at their
activity log, this adds a log entry in to the database so it can be
audited later.

This means each load of the page will trigger an `INSERT` which could
cause performance issues that we should monitor, and potentially look to
moving these records out of the main database in the future.