Handle sorting by year group with no registration
If the patient has no registration it can return `nil`, and we using
`try(:registration)` which also returns `nil` if the object doesn't
respond to that method. This lead to us thinking the object wasn't a
patient, and instead tried to call `.patient` on it which would then
fail. Instead we need to first check if the object responds to the
method before calling it.