User should have admin permissions if they're a superuser
This applies only in non-CIS2 environments (local development and
testing) to partially mimic the behaviour of CIS2 where a user has a
role and a superuser status that are independent from each other.
Currently in a non-CIS2 environment, the superuser status acts as its
own role, which means users with that status have very limited access
(they're able to access superuser actions only). To aid with testing, we
can assume that if the user is a superuser they should also have admin
permissions.

It's not currently possible to log in as a nurse user with superuser
permissions at the same time.