Handle generating consent forms for community clinics
In this case, the location would be the community clinic, but the user
would've been forced to choose a school so it would be different to the
location of the community clinic.

This should have no impact on the existing tests, but will fix an issue
in the `consent_forms:generate` Rake task.