Raise UnknownGPPractice if GP practice doesn't exist
Currently we raise an `ActiveRecord::RecordNotFound` error, but this
causes the job to be discarded by ActiveJob without an exception being
raised in Sentry as this kind of error can occur in other scenarios.

Instead, we can raise a specific error which is good practice anyway as
it ensures these get treated in a special way which allows us to get
a good idea of how often these happen anyway.