Handle situation where cis2 returns empty role
We had a Sentry error where our logging showed that in some scenarios
(we assume Smartcard native app use) CIS2 will send a user payload with
an empty selected role. Since we can't match their selected role to an
existing role and org, we should show an error in this scenario.