Update patient session links
This updates the links shown in the patient session table to link
directly to the patient within the session rather than linking to the
patient page which then requires the user to find them in the session.

This requires us to pick a suitable route in to the session according to
the status of the patient session, although this is likely to be less of
an issue when we implement the "Record session outcomes" page.