Add new patients to upcoming sessions
When resolving a duplicate patient by keeping both patients, we need to
ensure that the new patient is also included in any sessions that the
old patient was part of, otherwise the patient will end up in the global
children view but without any sessions.

This also ensures that any outstanding school moves for the old patient
are also copied to the new patient, this is necessary because a nurse
might resolve the duplicate before confirming the school move, so the
old patient might not be in the right session when it is duplicated.