Don't use not_invalidated scope
When determining the `latest_triage` for a patient session, and
therefore the status of a patient session. With a scope it forces the
query to re-run even if the triages were preloaded, whereas doing it in
Ruby means we avoid an extra query to the database. Because there are
unlikely to be lots of triages for a particular patient session, this
should improve performance.