Refactor component specs to use subject
This refactors the remaining component specs to use a `subject` instead of
a `before` block, allowing the use of `before` blocks to set things up
before the component is rendered.

This has been extracted from the strict loading work, where this was
needed to disable strict loading on a number of individual components
during testing in `before` blocks.