Use Patient#parent_relationships
This refactors the patient to avoid usage of the `parents` method and
instead use `parent_relationships` directly, which then avoids the need
to look up the relationship between the patient and the parent as it
will already be accessible.

This helps us to enable strict loading across the service by reducing
the number of queries that need to be made.