Prevent double-vaccination by navigating back
When using the vaccinate flow it's possible to inadvertantly record two
vaccinations for the same patient by navigating back to the confirmation
after submitting it and then submitting it again. Instead, once
recorded, we associate the draft vaccination record with the saved one
so this changes to the edit vaccination record flow.