Update PatientSession indexes
This updates the indexes on the patient session model to remove an
unnecessary unique index and add missing indexes on the patient and
session column.

The index on the individual columns should improve performance when
looking up individual patient sessions for a patient or a session,
related to the foreign key added in
b350c8eebae2097e008ee39f14db29d72ce55412.

The unique index on `session_id` and `patient_id` is unnecessary as
we already have an index on `patient_id` and `session_id`.