Pass parent relationship to AppParentSummaryComponent
This refactors the `AppParentSummaryComponent` class to pass in the
parent relationship directly, making it easier to enable strict loading
across the service by avoiding the need to call the `relationship_to`
method.

This is because using the `relationship_to` method requires additional
unnecessary queries whereas in all cases the relationship was already
available to use directly.