Choose an organisation on login without CIS2
We currently default to `current_user.organisations.first` if users
aren't signed in using CIS2 when deciding their `selected_organisation`.

This makes testing certain features harder, as there is no way for users
to sign into multiple organisations. Amongs other things, we could use
this capability in the training environment to set up 2 different
organisations; one that is a blank slate, one that has data in it, to
train on different scenarios.

This adds a page that simulates the "Select a role" page that users see
when they login via CIS2. However, it asks the users to choose an
organisation, not a role. The user is only asked on login (or during
their current session if they haven't chosen one already). In
environments where CIS2 is enabled, this functionality should not
affect anything.