Mark JavaScript tests with `:js`
This ensures that Capybara automatically uses the `javascript_driver`
and we don't need to mark the test as a system test. This avoids the
need to use `driven_by` and mark the test as a `:system` test while
being in the directory for feature tests.

In the future we might want to switch all these tests over to be
`:system` tests, as this is the standard Rails approach now, but for
now we can leave it like this.