Disable `:js` tests if Chrome is unavailable
We use Google Chrome (via Cuprite) to run feature tests with client-side
JavaScript. This updates the `spec_helper` to first detect whether Chrome
is available, and if not, disable `:js` tests. This is because not all
developers will have Chrome available on their machine.

We've only got one test that requires JavaScript so I thought this
would be a simpler approach for now, if we add more then we might want
to consider adding a Selenium WebDriver for Firefox or Safari as an
alternative.