Raise NHS::PDS::TooManyMatches
This updates the PDS search client code to handle getting a too many
matches error response and raise a suitable exception.

This doesn't handle the error, as we need to decide how to handle this
case, but having its own error means we can easily filter these out in
Sentry.