Add `birth_academic_year` column
This adds a new column to the patient model that will be used to track
the year of birth in relation to the academic year, allowing patients to
be placed in a year group that is not typical for their actual date of
birth. The same logic that we currently have on the `Cohort` model will
be applied to the `Patient` model to achieve this.