Set `birth_academic_year` column
This updates all code paths where a `Patient` instance is created to
ensure the `birth_academic_year` is set according to the `date_of_birth`
as the column is not nullable. Following this we will add the ability
for the nurses to customise this value by uploading a spreadsheet with a
column holding the year group if different to the one you would assume
from the year group.