Handle school move using `birth_academic_year`
This updates how a school move is confirmed to choose an approriate
`Cohort` according to the `birth_academic_year` value of the patient
rather than generating the academic year from the date of birth.

This is to ensure that if a patient is in a different year group for
what would be typical for their date of birth, we don't lose track of
that when they move schools. In the future we might want to provide the
nurses with an interface to choose what to do in this case.