Remove validation on year groups
When importing vaccination records we don't need to validate that the
year group is approriate for the current programme as we might be
importing historical records where the patient is no longer in the
cohort, but they did receive a vaccination at the time. We also don't
want to prevent real vaccinations from being imported where perhaps the
patient received their vaccination perhaps slightly earlier or later
than usual.