Allow specifying year group when importing
This allows the nurses to specify the year group of a particular patient
when importing them in via the cohort or class list import process. The
new column `CHILD_YEAR_GROUP` is optional, and if not provided we
fallback to the existing logic of calculating the year group from the
date of birth.