Fix sort/filter not working on attendances page
This fixes a bug that was inadvertently introduced in
ddbd6b96507cbf86388754fa72843c72adcc605d where users could no longer
sort and filter on the register attendances page. This was happening
because we try and call `select!` on an
`ActiveRecord::AssociationRelation` which doesn't have such a method,
instead we have to first convert the list of patients to an array.