Calculate `Consent#responded_at` from consent form
If the consent is linked to a consent form we should get the
`responded_at` date/time from when the consent form was submitted, not
when it was linked with the patient.

This has been spotted as a bug and the testers would like it included as
part of the work to fix inconsistent data between the consent and the
consent form
(https://github.com/nhsuk/manage-vaccinations-in-schools/pull/2772).