Move validation data to separate sheet for offline
In Excel spreadsheets when the list of available options for data
validation is too large this can cause issues.

https://stackoverflow.com/questions/16017403/maximum-drop-down-list-formula-length-in-excel

To avoid this we can move the list of options to another sheet, table or
named range.

https://contextures.com/xldataval11.html

This change moves the list of emails of the performing professionals to
it's own sheet and adds a formula to use them as the source data for the
data validation for the PERFORMING_PROFESSIONAL_EMAIL field of the
offline spreadsheet.

Fixes MAVIS-1800