Refactor AppOutcomeBannerComponent
This refactors the component to correctly handle the situation where a
patient was vaccinated in a session, and then a not administered
vaccination record was recorded _after_ the administered vaccination
record.

In reality this situation won't happen, but it can happen during
training, so we should update the component to handle this and avoid
presenting users with an internal server error.