Ensure `subsequent_session_dates_offered_message` is present
This is an optional personalisation variable that we send to GOV.UK
Notify and is used in the session reminder emails. Previously we were
not sending this variable when not present, but that was causing a bug
where the emails couldn't be sent as GOV.UK Notify would raise a client
bad request error when the template variable was missing. Instead we can
send it to GOV.UK Notify as a blank string.