Validate date of birth in past
When importing vaccination records we should validate that the date of
birth for the patient is in the past, it doesn't make sense for the date
to be in the future.

This validation was covered implicitly by the year group being in the
programme which was removed to handle customising the year groups of the
patients in
https://github.com/nhsuk/manage-vaccinations-in-schools/pull/2780 so
this is just a small amendment to that change to avoid a regression.