Avoid duplicate `DATE_OF_BIRTH` errors
This fixes a few minor visual issues spotted by the testers when testing
https://github.com/nhsuk/manage-vaccinations-in-schools/pull/2780 where
if the date of birth is missing the same error will appear twice for
each row, and similarly an issue where the year group wouldn't render
correctly.

This fixes are already in place for a future release (in
https://github.com/nhsuk/manage-vaccinations-in-schools/pull/2783 and
https://github.com/nhsuk/manage-vaccinations-in-schools/pull/2784),
however including them now makes the testing easier and avoids needing
to record additional defects.