Ignore restricted patients from sorting and filtering
This prevents information about restricted patients being leaked
inadvertently, as although the information itself was not shown, by
using the sorting and filtering functionality it's possible to guess
what the value might be.

Instead we now treat restricted patients as having no postcode for the
purpose of filtering and sorting.