Record who matched consent manually
When manually matching a consent form with a patient we should record
the user who performed the action, allowing us to show this information
in the activity log. It also provides a way of determining whether a
match was done automatically or manually, as automatically there will be
no user associated with the consent.