Improve email address and phone number normalisation
This improves how the email addresses and phone numbers are normalised
to make it easier to perform string comparisons on the values to match
up consent form email addresses and phone numbers with parent email
addresses and phone numbers. This is necessary to support the work to
match failed email/phone number attempts in Notify with consent forms,
where don't have a reference to the parent or patient instance.