Add NotifyDeliveryJob
This encapsulates some logic related to sending emails and texts via
GOV.UK Notify in to its own class so that in the future we can use this
when sending emails. We need to refactor the way emails are sent to use
the API directly so that we can integrate with the status receipts
provided by GOV.UK and this is the first step of that.