Add `delivery_status` column to `notify_log_entries`
This adds a new enum integer column to `notify_log_entries` that we can
use to track the status of a message delivery. This commit just adds the
column, but the next steps will be to integrate with the GOV.UK Notify
callback mechanism to ensure this value is accurate. This will allow us
to display this information to the user either via the UI or in a report.