Only consider triage status if consent is given
When calculating the overall status of a patient session, we should only
consider the status of the `latest_triage` if the consent is given.

This is to handle a scenario where consent was given, then the nurse
triaged the patient, and then consent was refused. Currently, the status
of the patient would remain "ready to vaccinate" whereas instead the
status should reflect the fact that consent has been withdrawn (even
though the patient was triaged at one point).