Create SchoolMoveLogEntry
This creates a new model that will be used to capture when a school move
is confirmed, by which user and outcome of the school move. I've only
added columns to capture the minimum we need to answer questions not
available in our existing audit logs, but this could be extended in the
future.

The model has been created using a similar pattern to other actions that
we log, specifically the `NotifyLogEntry` and `AccessLogEntry`. It might
be time to consider if we can build a more generic way of recording
actions taken in the service. For now, the approach taken in this commit
will allow us to answer the question of why a number of patients have
ended up in the community clinic when there are still school sessions
scheduled.

We already record database changes using the `audited` Gem, but this
doesn't always allow us to record actions taken in the service,
particular ones that aren't able to be determined by looking just at the
database changes, for example access logs, school moves or patient
merges. This is why it's been necessary to record these actions in
separate tables.